Refactor modulebuilder code
